moving all env variable configuration driven variables in arctic._config (#666)
* refactored moving all env variable configuration driven variables in the arctic._config module with comments describing its use

* change import to local dir

* fixed failing tests due to using STRICT_WRITE_HANDLER_MATCH instead of strict_write_handler
