+++ tests/integration/store/test_ndarray_store.py
@@ -6,7 +6,8 @@ from numpy.testing import assert_equal
 from pymongo.server_type import SERVER_TYPE
 import pytest
 
-from arctic._util import mongo_count, FwPointersCfg, FW_POINTERS_REFS_KEY
+from arctic._config import FwPointersCfg, FW_POINTERS_REFS_KEY
+from arctic._util import mongo_count
 from arctic.store._ndarray_store import NdarrayStore, _APPEND_COUNT
 from arctic.store.version_store import register_versioned_storage
